requirements.txt
fastapi
uvicorn
kaggle
python-dotenv
pydantic-settings
